requirements.txt
alpaca_trade_api==3.2.0
fastapi==0.115.12
joblib==1.1.0
numpy==1.21.5
pandas==1.4.4
pytest==8.3.5
python-dotenv==1.1.0
pytz==2022.1
scikit_learn==1.0.2
SQLAlchemy==1.4.39
